requirements.txt
datasets==3.5.0
huggingface_hub==0.30.2
keyring==25.6.0
PyPDF2==3.0.1
pytest==8.3.5
python-dotenv==1.1.0
Requests==2.32.3
urllib3==2.4.0
